+++ dbentry/site/views/edit.py
@@ -20,7 +20,7 @@ To declare inlines for handling relations:
 """
 
 from django import forms
-from django.urls import reverse_lazy
+from django.urls import reverse_lazy, reverse, NoReverseMatch
 
 from dbentry import models as _models
 from dbentry.autocomplete.widgets import make_widget
@@ -28,6 +28,7 @@ from dbentry.forms import GoogleBtnWidget
 from dbentry.site import forms as _forms
 from dbentry.site.registry import register_edit
 from dbentry.site.views.base import BaseEditView, Inline
+from dbentry.utils.url import urlname
 
 
 class BestandInline(Inline):
@@ -44,6 +45,7 @@ class AudioView(BaseEditView):
         model = _models.Audio.musiker.through
         verbose_model = _models.Musiker
         tabular = False
+        changelist_fk_field = "musiker"
 
     class BandInline(Inline):
         model = _models.Audio.band.through
@@ -162,7 +164,10 @@ class AusgabeView(BaseEditView):
     model = _models.Ausgabe
     fields = ["magazin", "status", "sonderausgabe", "e_datum", "jahrgang", "beschreibung", "bemerkungen"]
     inlines = [NumInline, MonatInline, LNumInline, JahrInline, AudioInline, VideoInline, BestandInline]
-    widgets = {"sonderausgabe": forms.Select(choices=[(True, "Ja"), (False, "Nein")])}
+    widgets = {
+        "sonderausgabe": forms.Select(choices=[(True, "Ja"), (False, "Nein")]),
+        "e_datum": forms.DateInput(attrs={"type": "date"}),
+    }
     require_confirmation = True
     confirmation_threshold = 0.8
 
@@ -417,6 +422,23 @@ class GenreView(BaseEditView):
     inlines = [AliasInline]
     require_confirmation = True
 
+    def get_changelist_links(self, labels=None):
+        links = super().get_changelist_links(labels)
+        # Add links to the Brochure models, if any.
+        # Links to these models are ignored by the default implementation in
+        # super() because the relation only exists on the BaseBrochure model
+        # (and not the concrete child models that "inherit" the relation)
+        # which does not have a changelist view and thus no URL.
+        for model in (_models.Brochure, _models.Kalender, _models.Katalog):
+            qs = model.objects.filter(genre=self.object)
+            if c := qs.count():
+                try:
+                    url = reverse(urlname("changelist", model._meta))
+                except NoReverseMatch:  # pragma: no cover
+                    continue
+                links.append((f"{url}?genre={self.object.pk}", model._meta.verbose_name_plural, c))
+        return links
+
 
 @register_edit(_models.Magazin)
 class MagazinView(BaseEditView):
@@ -569,6 +591,7 @@ class VideoView(BaseEditView):
         model = _models.Video.musiker.through
         verbose_model = _models.Musiker
         tabular = False
+        changelist_fk_field = "musiker"
 
     class BandInline(Inline):
         model = _models.Video.band.through