+++ tests/test_site/test_views.py
@@ -10,10 +10,11 @@ from django.contrib.admin.models import LogEntry
 from django.contrib.contenttypes.models import ContentType
 from django.http import HttpResponse
 from django.template import TemplateDoesNotExist
-from django.test import override_settings
-from django.urls import path, reverse
+from django.test import override_settings, TestCase
+from django.urls import path, reverse, NoReverseMatch
 from mizdb_tomselect.views import IS_POPUP_VAR
 
+from dbentry.site.forms import InlineForm
 from dbentry.site.views.base import (
     BaseEditView,
     Inline,
@@ -297,21 +298,21 @@ class TestBaseEditView(DataTestCase, ViewTestCase):
         view = self.get_view(self.get_request(), add=False)
         view.object = self.obj
         links = view.get_changelist_links()
-        self.assertIn((f"/musician/?band={self.obj.pk}", "Musicians (1)"), links)
+        self.assertIn((f"/musician/?band={self.obj.pk}", "Musicians", 1), links)
 
     def test_get_changelist_links_ignores_relations_with_inlines(self):
         """No changelist_links should be created for relations handled by inlines."""
         view = self.get_view(self.get_request(), add=False)
         view.object = self.obj
         links = view.get_changelist_links()
-        self.assertNotIn((f"/genre/?band={self.obj.pk}", "Genres (1)"), links)
+        self.assertNotIn((f"/genre/?band={self.obj.pk}", "Genres", 1), links)
 
     def test_get_changelist_links_prefer_labels_arg(self):
         """Passed in labels should be used over the model's verbose name."""
         view = self.get_view(self.get_request(), add=False)
         view.object = self.obj
         links = view.get_changelist_links(labels={"musician": "Hovercrafts"})
-        self.assertIn((f"/musician/?band={self.obj.pk}", "Hovercrafts (1)"), links)
+        self.assertIn((f"/musician/?band={self.obj.pk}", "Hovercrafts", 1), links)
 
     def test_get_changelist_links_uses_related_name(self):
         """If the relation has a related_name, it should be used as the label."""
@@ -320,7 +321,7 @@ class TestBaseEditView(DataTestCase, ViewTestCase):
             view = self.get_view(self.get_request(), add=False)
             view.object = self.obj
             links = view.get_changelist_links()
-            self.assertIn((f"/musician/?band={self.obj.pk}", "Hovercrafts Full Of Eels (1)"), links)
+            self.assertIn((f"/musician/?band={self.obj.pk}", "Hovercrafts Full Of Eels", 1), links)
 
     def test_initial_adds_preserved_filters(self):
         """
@@ -569,6 +570,17 @@ class TestBaseListView(DataTestCase, ViewTestCase):
         for i, expected_label in enumerate(expected_labels):
             with self.subTest(index=i, label=expected_label):
                 self.assertEqual(headers[i]["text"], expected_label)
+
+    def test_get_result_headers_field_not_in_sortable_by(self):
+        """
+        Assert that a header is flagged as not sortable if the field is not
+        included in the view's ``sortable_by``.
+        """
+        view = self.get_view(self.get_request())
+        view.sortable_by = ["name"]
+        headers = view.get_result_headers()
+        self.assertFalse(headers[1].get("sortable", False))
+        self.assertTrue(headers[0].get("sortable", False))  # name should still be sortable
 
     def test_get_result_headers_no_list_display(self):
         """
@@ -962,6 +974,19 @@ class TestBaseListView(DataTestCase, ViewTestCase):
         self.assertMessageSent(response.wsgi_request, "Abgebrochen")
         self.assertRedirects(response, self.url)
 
+    def test_post_export_results(self):
+        """Assert that the export_results action is called with the view's queryset."""
+        request = self.post_request(data={"action_name": "export_results"})
+        view = self.get_view(request=request)
+        queryset_mock = Mock()
+        export_results_mock = Mock()
+        export_results_mock.__name__ = "export_results"
+        with patch.object(view, "get_queryset", new=Mock(return_value=queryset_mock)):
+            with patch.object(view, "get_actions") as get_actions_mock:
+                get_actions_mock.return_value = {"export_results": (export_results_mock, "foo", "bar")}
+                view.post(request)
+                export_results_mock.assert_called_with(view, request, queryset_mock)
+
 
 @override_settings(ROOT_URLCONF=URLConf)
 class TestDeleteView(ViewTestCase):
@@ -970,8 +995,8 @@ class TestDeleteView(ViewTestCase):
 
     @classmethod
     def setUpTestData(cls):
-        cls.obj = make(Country)
-        cls.band = make(Band, origin=cls.obj)
+        cls.obj = obj = make(Country)
+        cls.band = make(Band, origin=obj)
         super().setUpTestData()
 
     def setUp(self):
@@ -1089,8 +1114,8 @@ class TestSearchableListView(ViewTestCase):
     @classmethod
     def setUpTestData(cls):
         super().setUpTestData()
-        cls.fighters = make(Band, name="Foo Fighters")
-        cls.obj = make(cls.model, name="Foo Fighter", band=cls.fighters)
+        cls.fighters = fighters = make(Band, name="Foo Fighters")
+        cls.obj = make(cls.model, name="Foo Fighter", band=fighters)
 
     def test_get_search_results(self):
         """Assert that get_search_results returns the expected queryset result."""
@@ -1190,3 +1215,132 @@ class TestHelpView(ViewTestCase):
         """
         response = self.client.get(reverse("help", kwargs={"page_name": "__foo__"}), follow=True)
         self.assertMessageSent(response.wsgi_request, "Hilfe Seite für '__foo__' nicht gefunden.")
+
+
+@override_settings(ROOT_URLCONF=URLConf)
+class TestInline(TestCase):
+    class GenreInline(Inline):
+        model = Band.genres.through
+        verbose_model = Genre
+
+    def test_get_name_verbose_name_set(self):
+        with patch.object(self.GenreInline, "verbose_name", new="Foo"):
+            inline = self.GenreInline(Band)
+            self.assertEqual(inline._get_name("verbose_name"), "Foo")
+
+    def test_get_name_verbose_name_plural_set(self):
+        with patch.object(self.GenreInline, "verbose_name_plural", new="Foo Plural"):
+            inline = self.GenreInline(Band)
+            self.assertEqual(inline._get_name("verbose_name_plural"), "Foo Plural")
+
+    def test_get_name_verbose_model_set(self):
+        with patch.object(self.GenreInline, "verbose_model", new=Genre):
+            inline = self.GenreInline(Band)
+            self.assertEqual(inline._get_name("verbose_name"), Genre._meta.verbose_name)
+            self.assertEqual(inline._get_name("verbose_name_plural"), Genre._meta.verbose_name_plural)
+
+    def test_get_no_attr_or_verbose_model_set(self):
+        with patch.object(self.GenreInline, "verbose_model", new=None):
+            inline = self.GenreInline(Band)
+            self.assertEqual(inline._get_name("verbose_name"), inline.model._meta.verbose_name)
+            self.assertEqual(inline._get_name("verbose_name_plural"), inline.model._meta.verbose_name_plural)
+
+    def test_init_sets_verbose_names(self):
+        inline = self.GenreInline(Band)
+        self.assertEqual(inline.verbose_name, Genre._meta.verbose_name)
+        self.assertEqual(inline.verbose_name_plural, Genre._meta.verbose_name_plural)
+
+    def test_get_formset_class(self):
+        inline = self.GenreInline(Band)
+        formset_class = inline.get_formset_class()
+        self.assertEqual(formset_class.model, inline.model)
+        self.assertEqual(formset_class.fk.related_model, Band)
+        self.assertTrue(issubclass(formset_class.form, InlineForm))
+        self.assertEqual(formset_class.extra, 1)
+
+    def test_get_changelist_url_attr_set(self):
+        """
+        Assert that ``get_changelist_url`` returns the value of the
+        `changelist_url` attribute if it is set.
+        """
+        inline = self.GenreInline(Band)
+        inline.changelist_url = "foo"
+        with patch.object(inline, "get_changelist_fk_field", Mock(return_value="genre")):
+            self.assertEqual(inline.get_changelist_url(), "foo")
+
+    def test_get_changelist_url_id_field_set(self):
+        """
+        Assert that ``get_changelist_url`` returns the expected URL if a
+        changelist_fk_field could be found.
+        """
+        inline = self.GenreInline(Band)
+        with patch.object(inline, "get_changelist_fk_field", Mock(return_value="genre")):
+            self.assertEqual(inline.get_changelist_url(), "/genre/")
+
+    def test_get_changelist_url_no_reverse_match(self):
+        """
+        Assert that ``get_changelist_url`` returns an empty string if the URL
+        could not be reversed.
+        """
+        inline = self.GenreInline(Band)
+        with patch.object(inline, "get_changelist_fk_field", Mock(return_value="genre")):
+            with patch("dbentry.site.views.base.reverse") as reverse_mock:
+                reverse_mock.side_effect = NoReverseMatch
+                self.assertEqual(inline.get_changelist_url(), "")
+
+    def test_get_changelist_url_field_is_none(self):
+        """
+        Assert that ``get_changelist_url`` returns an empty string if
+        get_changelist_fk_field returns None.
+        """
+        inline = self.GenreInline(Band)
+        with patch.object(inline, "get_changelist_fk_field", Mock(return_value=None)):
+            self.assertEqual(inline.get_changelist_url(), "")
+
+    def test_get_changelist_fk_field_none(self):
+        """
+        Assert that ``get_changelist_fk_field`` returns an empty string if
+        changelist_fk_field is set to None.
+        """
+        inline = self.GenreInline(Band)
+        inline.changelist_fk_field = None
+        self.assertEqual(inline.get_changelist_fk_field(), "")
+
+    def test_get_changelist_fk_field_attr_set(self):
+        """
+        Assert that ``get_changelist_fk_field`` returns the value of the
+        changelist_fk_field attribute if it is set.
+        """
+        inline = self.GenreInline(Band)
+        inline.changelist_fk_field = "foo"
+        self.assertEqual(inline.get_changelist_fk_field(), "foo")
+
+    def test_get_changelist_fk_field_more_than_three_fields(self):
+        """
+        Assert that ``get_changelist_fk_field`` returns an empty string if the
+        inline model has more than three model fields.
+        """
+        inline = self.GenreInline(Band)
+        with patch.object(inline.model._meta, "get_fields", Mock(return_value=[1, 2, 3, 4])):
+            self.assertEqual(inline.get_changelist_fk_field(), "")
+
+    def test_get_changelist_fk_field(self):
+        """
+        Assert that ``get_changelist_fk_field`` returns the name of the
+        ForeignKey field to the target model.
+        """
+        inline = self.GenreInline(Band)
+        self.assertEqual(inline.get_changelist_fk_field(), "genre")
+
+    def test_get_context_data(self):
+        inline = self.GenreInline(Band)
+        expected = {
+            "verbose_name": "Genre",
+            "verbose_name_plural": "Genres",
+            "model_name": "band_genres",
+            "add_text": "Genre hinzufügen",
+            "tabular": True,
+            "changelist_url": "/genre/",
+            "changelist_fk_field": "genre",
+        }
+        self.assertEqual(inline.get_context_data(), expected)